+++ eras/shelley/impl/test/Test/Cardano/Ledger/Shelley/Binary/CddlSpec.hs
@@ -23,6 +23,7 @@ import Test.Cardano.Ledger.Binary.Cuddle (
   huddleRoundTripAnnCborSpec,
   huddleRoundTripArbitraryValidate,
   huddleRoundTripCborSpec,
+  noTwiddle,
   specWithHuddle,
  )
 import Test.Cardano.Ledger.Common
@@ -32,7 +33,7 @@ spec :: Spec
 spec =
   describe "CDDL" $ do
     let v = eraProtVerLow @ShelleyEra
-    describe "Huddle" $ specWithHuddle shelleyCDDL $ do
+    describe "Huddle" . specWithHuddle shelleyCDDL . noTwiddle $ do
       huddleRoundTripCborSpec @Addr v "address"
       huddleRoundTripArbitraryValidate @Addr v "address"
       huddleRoundTripAnnCborSpec @BootstrapWitness v "bootstrap_witness"