+++ libs/cardano-protocol-tpraos/test/Test/Cardano/Protocol/Binary/CddlSpec.hs
@@ -42,7 +42,7 @@ specForEra ::
 specForEra cddlFiles = do
   describe (eraName @era) $ do
     describe "Huddle" $
-      specWithHuddle cddlFiles $ do
+      specWithHuddle cddlFiles . noTwiddle $ do
         huddleBlockSpec @era @StandardCrypto @BHeader @BHBody
         xdescribe "Cannot generate a CBOR term corresponding to a group with cuddle" $
           huddleRoundTripCborSpec @(CBORGroup (OCert StandardCrypto))