+++ eras/shelley/impl/CHANGELOG.md
@@ -40,6 +40,7 @@
 
 ### `testlib`
 
+* Add `HuddleRule "int32" ShelleyEra` instance
 * Add `withIssuerAndTxsInBlock_` and `withIssuerAndTxsInBlock`
 * Add a `Maybe (KeyHash BlockIssuer)` parameter to `withTxsInBlockEither`